backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
aiofiles==23.1.0
jinja2==3.1.2

